reorganize functions and files
