+++ mcx/inference/warmup/step_size_adaptation.py
@@ -23,13 +23,11 @@ from mcx.inference.kernels import HMCState
 __all__ = [
     "dual_averaging",
     "find_reasonable_step_size",
-    "mass_matrix_adaptation",
-    "longest_batch_before_turn",
 ]
 
 
 # --------------------------------------
-#      == STEP SIZE ADAPTATION ==
+#      == DUAL AVERAGING ==
 # --------------------------------------
 
 
@@ -141,6 +139,11 @@ def dual_averaging(
     return init, update
 
 
+# --------------------------------------------
+#      == FIND REASONABLE INITIAL VALUE ==
+# --------------------------------------------
+
+
 def find_reasonable_step_size(
     rng_key: jax.random.PRNGKey,
     kernel_generator: Callable[[float], Callable],
@@ -208,201 +211,3 @@ def find_reasonable_step_size(
         _do_continue, _update, (rng_key, 0, 0, initial_step_size)
     )
     return step_size
-
-
-# --------------------------------------
-#      == MASS MATRIX ADAPTATION ==
-# --------------------------------------
-
-
-class WelfordAlgorithmState(NamedTuple):
-    mean: float  # the current sample mean
-    m2: float  # the current value of the sum of difference of squares
-    sample_size: int  # the sample size
-
-
-class MassMatrixAdaptationState(NamedTuple):
-    inverse_mass_matrix: np.DeviceArray
-    wc_state: WelfordAlgorithmState
-
-
-def mass_matrix_adaptation(
-    is_diagonal_matrix: bool = True,
-) -> Tuple[Callable, Callable, Callable]:
-    """Adapts the values in the mass matrix by computing the covariance
-    between parameters.
-    """
-    wc_init, wc_update, wc_final = welford_algorithm(is_diagonal_matrix)
-
-    def init(n_dims: int) -> MassMatrixAdaptationState:
-        if is_diagonal_matrix:
-            inverse_mass_matrix = np.ones(n_dims)
-        else:
-            inverse_mass_matrix = np.identity(n_dims)
-
-        wc_state = wc_init(n_dims)
-
-        return MassMatrixAdaptationState(inverse_mass_matrix, wc_state)
-
-    def update(
-        state: MassMatrixAdaptationState, value: np.DeviceArray
-    ) -> MassMatrixAdaptationState:
-        inverse_mass_matrix, wc_state = state
-        wc_state = wc_update(wc_state, value)
-        return MassMatrixAdaptationState(inverse_mass_matrix, wc_state)
-
-    @partial(jax.jit, static_argnums=(0,))
-    def final(state: MassMatrixAdaptationState) -> MassMatrixAdaptationState:
-        _, wc_state = state
-        covariance, count, mean = wc_final(wc_state)
-
-        # Regularize the covariance matrix, see Stan
-        scaled_covariance = (count / (count + 5)) * covariance
-        shrinkage = 1e-3 * (5 / (count + 5))
-        if is_diagonal_matrix:
-            inverse_mass_matrix = scaled_covariance + shrinkage
-        else:
-            inverse_mass_matrix = scaled_covariance + shrinkage * np.identity(
-                mean.shape[0]
-            )
-
-        ndims = np.shape(inverse_mass_matrix)[-1]
-        new_mm_state = MassMatrixAdaptationState(inverse_mass_matrix, wc_init(ndims))
-
-        return new_mm_state
-
-    return init, update, final
-
-
-def welford_algorithm(is_diagonal_matrix: bool) -> Tuple[Callable, Callable, Callable]:
-    """Welford's online estimator of covariance.
-
-    It is possible to compute the variance of a population of values in an
-    on-line fashion to avoid storing intermediate results. The naive recurrence
-    relations between the sample mean and variance at a step and the next are
-    however not numerically stable.
-
-    Welford's algorithm uses the sum of square of differences
-    :math:`M_{2,n} = \\sum_{i=1}^n \\left(x_i-\\overline{x_n}\right)^2`
-    for updating where :math:`x_n` is the current mean and the following
-    recurrence relationships
-
-    .. math:
-        M_{2,n} = M_{2, n-1} + (x_n-\\overline{x}_{n-1})(x_n-\\overline{x}_n)
-        \\sigma_n^2 = \\frac{M_{2,n}}{n}
-    """
-
-    def init(n_dims: int) -> WelfordAlgorithmState:
-        """Initialize the covariance estimation.
-
-        When the matrix is diagonal it is sufficient to work with an array that contains
-        the diagonal value. Otherwise we need to work with the matrix in full.
-
-        Argument
-        --------
-        n_dims: int
-            The number of dimensions of the problem, which corresponds to the size
-            of the corresponding square mass matrix.
-        """
-        count = 0
-        mean = np.zeros(n_dims)
-        if is_diagonal_matrix:
-            m2 = np.zeros(n_dims)
-        else:
-            m2 = np.zeros((n_dims, n_dims))
-        return WelfordAlgorithmState(mean, m2, count)
-
-    @jax.jit
-    def update(
-        state: WelfordAlgorithmState, value: np.DeviceArray
-    ) -> WelfordAlgorithmState:
-        """Update the M2 matrix using the new value.
-
-        Arguments:
-        ----------
-        state:
-            The current state of the Welford Algorithm
-        value: jax.numpy.DeviceArray, shape (1,)
-            The new sample used to update m2
-        """
-        mean, m2, count = state
-        count = count + 1
-
-        delta = value - mean
-        mean = mean + delta / count
-        updated_delta = value - mean
-        if is_diagonal_matrix:
-            new_m2 = m2 + delta * updated_delta
-        else:
-            new_m2 = m2 + np.outer(delta, updated_delta)
-
-        return WelfordAlgorithmState(mean, new_m2, count)
-
-    def covariance(state: WelfordAlgorithmState) -> np.DeviceArray:
-        mean, m2, count = state
-        covariance = m2 / (count - 1)
-        return covariance, count, mean
-
-    return init, update, covariance
-
-
-# ------------------------------------------
-#      == INTEGRATION STEPS ADAPTATION ==
-# ------------------------------------------
-
-
-def longest_batch_before_turn(integrator_step: Callable) -> Callable:
-    """Learn the number of steps one can make before the trajectory makes a
-    U-Turn. This routine is part of the adaptive strategy described in [1]_:
-    during the warmup phase we run this scheme many times in order to get a
-    distribution of numbers of steps before U-Turn. We then sample elements
-    from this distribution during inference to use as the number of integration
-    steps.
-
-    References
-    ----------
-    .. [1]: Wu, Changye, Julien Stoehr, and Christian P. Robert. "Faster
-            Hamiltonian Monte Carlo by learning leapfrog scale." arXiv preprint
-            arXiv:1810.04449 (2018).
-    """
-
-    @partial(jax.jit, static_argnums=(0, 1, 2, 3))
-    def run(
-        initial_position: np.DeviceArray,
-        initial_momentum: np.DeviceArray,
-        step_size: float,
-        num_integration_steps: int,
-    ):
-        def cond(state: Tuple) -> bool:
-            iteration, position, momentum = state
-            return is_u_turn or iteration == num_integration_steps
-
-        def update(state: Tuple) -> Tuple:
-            iteration, position, momentum = state
-            iteration += 1
-            position, momentum = integrator_step(position, momentum, step_size, 1)
-            return (iteration, position, momentum)
-
-        result = jax.lax.while_loop(
-            cond, update, (0, initial_position, initial_momentum)
-        )
-
-        return result[0]
-
-    return run
-
-
-@partial(jax.jit, static_argnums=(0, 2))
-def is_u_turn(
-    initial_position: np.DeviceArray,
-    position: np.DeviceArray,
-    inverse_mass_matrix: np.DeviceArray,
-    momentum: np.DeviceArray,
-) -> bool:
-    """Detect when the trajectory starts turning back towards the point
-    where it started.
-    """
-    v = np.multiply(inverse_mass_matrix, momentum)
-    position_vec = position - initial_position
-    projection = np.multiply(position_vec, v)
-    return np.where(projection < 0, True, False)
